you can stop the cronjob now
