+++ src/types/alarms.ts
@@ -1,0 +1,11 @@
+interface AlarmThresholds {
+  timeThreshold: number;
+  storageThreshold: number;
+  reminderInterval: number;
+}
+
+export interface Alarm {
+  id: string;
+  data: AlarmThresholds;
+}
+
